refactor NamedFilter
